refactor(logic): split quick battle resolver into library parts (Refs #2288)
Extract round-engine helpers and outcome/finish paths into part files so the
main resolver stays under the monolithic file threshold. No behavior change;
same private API within the quick_battle_resolver library.

Deferred: remaining >700-line production files and test splits per #2288.
